add basic logger

+++ apple/xray/Core/Events/Event.swift
@@ -11,16 +11,16 @@ import Foundation
 class Event: NSObject {
     let category: String?
     let subsystem: String
-    let timestamp: Double // UTC
+    let timestamp: UInt // UTC
     let level: Int
     let message: String
     let data: [String: Any]? // What func, line
     let context: [String: Any]?
-    let exception: NSException? // todo: not sure about this one, maybe handle it in event builder and add to data?
+    let exception: NSException?
 
     init(category: String?,
          subsystem: String,
-         timestamp: Double,
+         timestamp: UInt,
          level: Int = 0,
          message: String,
          data: [String: Any]?,
